// backend/src/enums/finance.enums.ts
export enum AccountType {
  CHECKING = "CHECKING",
  SAVINGS = "SAVINGS",
  CASH = "CASH",
  INVESTMENT = "INVESTMENT",
  CREDIT_CARD = "CREDIT_CARD",
}

export enum CategoryType {
  INCOME = "INCOME",
  EXPENSE = "EXPENSE",
  BOTH = "BOTH",
}

export enum TransactionType {
  INCOME = "INCOME",
  EXPENSE = "EXPENSE",
  TRANSFER = "TRANSFER",
  ADJUSTMENT = "ADJUSTMENT",
}

export enum AssetType {
  STOCK = "STOCK",
  FII = "FII",
  ETF = "ETF",
  BDR = "BDR",
  REIT = "REIT",
  CRYPTO = "CRYPTO",
  FIXED_INCOME = "FIXED_INCOME",
  CASH_BOX = "CASH_BOX",
  OTHER = "OTHER",
}

export enum InvestmentOperationType {
  BUY = "BUY",
  SELL = "SELL",
  DIVIDEND = "DIVIDEND",
  INTEREST = "INTEREST",
  DEPOSIT = "DEPOSIT",
  WITHDRAW = "WITHDRAW",
}